requirements.txt
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-openai==0.2.10
azure-search-documents==11.5.2
openai==1.55.1
SQLAlchemy==2.0.35
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
